refactor: `HttpError` 관련 class, handle 함수 디렉터리 위치 변경
